+++ src/main/kotlin/com/draculatheme/jetbrains/settings/DraculaState.kt
@@ -1,0 +1,5 @@
+package com.draculatheme.jetbrains.settings
+
+class DraculaState {
+    var version: String = ""
+}